requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
et-xmlfile==1.1.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
lxml==4.9.2
msal==1.22.0
multitasking==0.0.11
numpy==1.24.3
Office365-REST-Python-Client==2.4.1
openpyxl==3.1.2
pandas==2.0.2
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.2
webencodings==0.5.1
yfinance==0.2.18
